Cache improved

+++ YtDlpViewsDumper.py
@@ -12,6 +12,7 @@ from datetime import datetime
 import yt_dlp
 import os
 import asyncio
+import concurrent
 from concurrent.futures import ThreadPoolExecutor
 import atexit
 import logging
@@ -47,9 +48,18 @@ class CacheManager:
         cache_dir_local = os.path.join(self.cache_dir_global, cache_name)
         return cache_dir_local
     
+    def get_cache_creation_time(self, cache_name):
+        cache_dir = self.get_cache_dir_local(cache_name)
+
+        if not os.path.exists(cache_dir):
+            return None
+
+        return os.path.getctime(cache_dir)
+       
     def clear_cache(self, cache_name):
         path = self.get_cache_dir_local(cache_name)
-        shutil.rmtree(path)
+        if os.path.exists(path):
+            shutil.rmtree(path)
     
     def append(self, cache_name, records: dict | list[dict], chunk_size_override = None):
         if chunk_size_override:
@@ -65,11 +75,11 @@ class CacheManager:
             os.makedirs(cache_dir)
             logger.debug(f"Created cache dir for {cache_name}")
         
-        if not self.cache_buffer[cache_name]:
+        if cache_name not in self.cache_buffer:
             self.cache_buffer[cache_name] = []
 
         for record in records:
-            if len(self.cache_buffer[cache_name]) > chunk_size_local:
+            if len(self.cache_buffer[cache_name]) >= chunk_size_local:
                 self.flush_buffers()
                 
             self.cache_buffer[cache_name].append(record)
@@ -80,7 +90,7 @@ class CacheManager:
         cache_dir = self.get_cache_dir_local(cache_name)
         if not os.path.isdir(cache_dir):
             logger.debug(f"Cache {cache_name} does not exist")
-            return None
+            return []
 
         for file_path_relative in os.listdir(cache_dir):
             file_path = os.path.join(cache_dir, file_path_relative)
@@ -101,25 +111,27 @@ class CacheManager:
             logger.debug(f"Flushing {len(buffer)} buffered records for cache {cache_name}")            
             cache_dir = self.get_cache_dir_local(cache_name)
             chunk_names = os.listdir(cache_dir)
-            chunk_max = 0
+            chunk_max = None
             for chunk_name in chunk_names:
-                chunk_num_str = chunk_name[len(cache_name) + 1:len(".cache.json")]
+                chunk_num_str = chunk_name[len(cache_name) + 1 : len(chunk_name) - len(".cache.json")]
                 chunk_num = int(chunk_num_str)
-                if chunk_max < chunk_num:
+                if chunk_max is None or chunk_max < chunk_num:
                     chunk_max = chunk_num
                     
             logger.debug(f"Max existing chunk number for {cache_name}: {chunk_max}")
-            chunk_num_current = chunk_max + 1
-            cache_chunk_name = os.path.join(cache_dir, f"{cache_name}-{chunk_num_current:6}.cache.json")
+            if chunk_max is None:
+                chunk_num_current = 0
+            else:
+                chunk_num_current = chunk_max + 1
+            cache_chunk_name = os.path.join(cache_dir, f"{cache_name}-{chunk_num_current:06}.cache.json")
             
-            with open(cache_chunk_name) as fout:
+            with open(cache_chunk_name, "w") as fout:
                 json.dump(buffer, fout, indent=4)
             
             logger.debug(f"Flushed chunk {chunk_num_current} with {len(buffer)} records to {cache_chunk_name}")
 
             buffer.clear()
 
-cache_manager = None    
 
 def get_video_metadata(video, credentials = None, yt_dlp_path = "yt-dlp"):
     
@@ -204,6 +216,7 @@ def get_cache_creation_time(cache_dir, cache_name):
 
     return os.path.getctime(cache_file)
 
+'''
 def read_cache(cache_dir, cache_name):
     if not cache_dir:
         cache_dir = os.getcwd()
@@ -223,7 +236,7 @@ def write_cache(cache_dir, cache_name, data):
 
     with open(cache_file, "w") as outfile:
         json.dump(data, outfile, indent=4)
-
+'''
 
 def generate_periodical_ticks(timestamp_start, timestamp_end, period_seconds = 60*60*24*30, max_ticks = 25):
     xticks_values = []
@@ -498,114 +511,121 @@ def fetch_metadata(video):
         logger.error(f"Failed to get metadata for {video}: {ex}")
         return None
 
+async def fetch_channel_data_offline(channel, cache_manager):
+    logger.info("Reading from cache (offline mode)...")
+    username = get_username_from_url(channel)
 
-async def fetch_channel_data(channel, cache_dir, yt_dlp, jobs, date_from_seconds, cache_expiration_seconds, offline):
+    dataset = cache_manager.read_cache(username)
+    logger.info(f"Read {len(dataset)} videos from cache {username}")
+    
+    logger.info(f"Done for {channel}!")
+    return dataset, username
+
+def remove_outdated_entries(cached_dataset, cache_manager, timestamp_current, username, cache_expiration_seconds):
+    cache_outdated_entries = []
+    for cached_entry in cached_dataset:
+        cache_timestamp = cached_entry["cache_timestamp"] if "cache_timestamp" in cached_entry else None
+        
+        if cache_timestamp is None:
+            cache_timestamp = cache_manager.get_cache_creation_time(username)
+        
+        if cache_expiration_seconds and \
+            cache_expiration_seconds > 0 and \
+            timestamp_current - cache_timestamp > cache_expiration_seconds:
+            cache_outdated_entries.append(i)
+    logger.info(f"{len(cache_outdated_entries)} entries in cache {username} are outdated")
+    
+    for entry in cache_outdated_entries:
+        cached_dataset.remove(entry)
+
+async def fetch_channel_data(channel, cache_dir, yt_dlp, jobs, date_from_seconds, cache_expiration_seconds, cache_manager, offline):
     dataset = []
 
     timestamp_current = datetime.now().timestamp()
 
     username = get_username_from_url(channel)
 
-    if offline and cache_exists(cache_dir, username):
-        logger.info("Reading from cache...")
-        username = get_username_from_url(channel)
-
-        dataset = read_cache(cache_dir, username)
-    else:
-        outdated_cache_num = 0
-
-        logger.info(f"Downloading video list from {channel}...")
-        videos = get_video_list(channel, yt_dlp)
-        logger.info("Done!")
-
-        videos_num = len(videos)
-        logger.info(f"{videos_num} videos found!")
-
-        logger.info("Reading from cache and downloading...")
-        username = get_username_from_url(channel)
-
-        cached_dataset = None
+    if offline:
         if cache_exists(cache_dir, username):
-            cached_dataset = read_cache(cache_dir, username)
+            return await fetch_channel_data_offline(channel, cache_dir)
+        else:
+            logger.error("Cache {} does not exist, impossible to work in offline mode!")
+            return None, username
 
-        with alive_bar(videos_num, title="Downloading metadata", theme="classic", force_tty=True, title_length=0) as bar:
-            videos_to_load = []
-            
-            for video in videos:
-                must_download = True
-                cached_entry = find_by_url(cached_dataset, video)
-                if cached_entry is not None:
-                    cache_timestamp = cached_entry["cache_timestamp"] if "cache_timestamp" in cached_entry else None
-                    
-                    if cache_timestamp is None:
-                        # cache_timestamp = datetime.datetime.now().timestamp()
-                        cache_timestamp = get_cache_creation_time(cache_dir, username)
-                    
-                    if cache_expiration_seconds and \
-                        cache_expiration_seconds > 0 and \
-                        timestamp_current - cache_timestamp > cache_expiration_seconds:
+    outdated_cache_num = 0
 
-                        outdated_cache_num += 1
-                    else:
-                        dataset.append(cached_entry)
-                        bar(1, skipped=True)
-                        must_download = False
+    logger.info(f"Downloading video list from {channel}...")
+    videos = get_video_list(channel, yt_dlp)
 
-                if must_download:
-                    videos_to_load.append(video)
-            
-            if outdated_cache_num > 0:
-                logger.info(f"{outdated_cache_num} cache entries were outdated")
+    videos_num = len(videos)
+    logger.info(f"{videos_num} videos found!")
 
-            futures = []
-            with ThreadPoolExecutor(max_workers=jobs) as executor:
+    logger.info("Reading from cache and downloading...")
 
-                for video in videos_to_load:
-                    future = executor.submit(fetch_metadata, video)
-                    futures.append(future)
-                    
-                while True:
-                    has_active_futures = False
+    cached_dataset = cache_manager.read_cache(username)
+    if cached_dataset:
+        logger.info(f"Read {len(cached_dataset)} videos from cache {username}")
+    else:
+        logger.info(f"Cache {username} does not exist or is empty.")
+    
+    remove_outdated_entries(cached_dataset, cache_manager, timestamp_current, username, cache_expiration_seconds)
+    logger.info(f"Rewriting cache {username} with outdated elements removed...")
+    cache_manager.clear_cache(username)
+    cache_manager.append(username, cached_dataset, len(cached_dataset))
+    cache_manager.flush_buffers()
 
-                    done_futures = []
+    dataset += cached_dataset
 
-                    for future in futures:
-                        if not future.done():
-                            has_active_futures = True
-                            continue
-                        if not future.cancelled():
-                            done_futures.append(future)
+    with alive_bar(videos_num, title="Downloading metadata", theme="classic", force_tty=True, title_length=0) as bar:
+        videos_to_load = []
+        
+        for video in videos:
+            cached_entry = find_by_url(cached_dataset, video)
+            if cached_entry is None:
+                videos_to_load.append(video)
+            else:
+                bar(1, skipped=True)
+        
+        futures = []
+        with ThreadPoolExecutor(max_workers=jobs) as executor:
 
-                    for future in done_futures:
+            for video in videos_to_load:
+                future = executor.submit(fetch_metadata, video)
+                futures.append(future)
+                
+            while futures:
+                concurrent.futures.wait(futures, timeout=None, return_when="FIRST_COMPLETED")
+                done_futures = concurrent.futures.as_completed(futures, timeout=None)
+
+                for future in done_futures:
+                    try:
                         metadata = future.result()
-                        if metadata:
-                            metadata["cache_timestamp"] = datetime.now().timestamp()
-                            dataset.append(metadata)
-                            write_cache(cache_dir, username, dataset)
+                        metadata["cache_timestamp"] = datetime.now().timestamp()
+                        dataset.append(metadata)
+                        # write_cache(cache_dir, username, dataset)
+                        cache_manager.append(username, metadata)
+                    
+                        upload_date_seconds = metadata["date"]
+                        upload_date = datetime.fromtimestamp(upload_date_seconds)
+
+                        if date_from_seconds and upload_date_seconds < date_from_seconds:
+                            logger.info("Minimum timestamp reached")
+                            future_index = futures.index(future)
+
+                            if len(futures) > future_index:
+                                discarded_futures = futures[future_index + 1 : -1]
+                                for i, discard_future in enumerate(discarded_futures):
+                                    discard_future.cancel()
+                    except Exception as err:
+                        logger.error(f"Downloading error: {err}")
                         
-                            upload_date_seconds = metadata["date"]
-                            upload_date = datetime.fromtimestamp(upload_date_seconds)
+                    futures.remove(future)
 
-                            if date_from_seconds and upload_date_seconds < date_from_seconds:
-                                logger.info("Minimum timestamp reached")
-                                future_index = futures.index(future)
-
-                                if len(futures) > future_index:
-                                    discarded_futures = futures[future_index + 1 : -1]
-                                    for i, discard_future in enumerate(discarded_futures):
-                                        discard_future.cancel()
-
-                        futures.remove(future)
-
-                        bar.title(upload_date.strftime("%d.%m.%Y"))
-                        bar(1)
-
-                    if not has_active_futures:
-                        break
-
-                    await asyncio.sleep(0.25)
+                    bar.title(upload_date.strftime("%d.%m.%Y"))
+                    bar(1)
 
     logger.info(f"Done for {channel}!")
+    cache_manager.flush_buffers()
     return dataset, username
 
 
@@ -686,7 +706,7 @@ async def main():
     usernames = []
 
     for channel in args.channels:
-        dataset, username = await fetch_channel_data(channel, cache_dir, args.yt_dlp, args.jobs, date_from_seconds, cache_expiration_seconds, args.offline)
+        dataset, username = await fetch_channel_data(channel, cache_dir, args.yt_dlp, args.jobs, date_from_seconds, cache_expiration_seconds, cache_manager, args.offline)
         channel_data_dict[channel] = (dataset, username)
         usernames.append(username)
